Add dedicated functions to back up the commands, so that error messages are printed
Add Concealer#List() and implement this for both global and local conceal groups (in different ways, due to the ways they are stored).
Adapt s:SetConcealDefaults() to work on global conceal groups: Invoking on the first window of each buffer isn't sufficient, as it sets window-scoped stuff. Rather, apply once for local conceal groups, and for each window (repeated once when each tab page is entered) for global conceal groups.