FIX: Need to set b:Concealer_Local_Chars in Conceal#AddLocal()
As that is advertised as an API function, and can be passed a custom char as well. So let's move the initialization from Concealer#Here() there, and add corresponding cleanup in Concealer#RemoveLocal().
As the custom char cannot be clearly discerned in Concealer#AddLocal() (its not an optional a:1 argument there), we need to actually check whether the key is a count (then it's a configured char), or not (then it's a custom char). This follows the logic in s:GetLocalChar() and s:EchoConceal().
(If the user mistakenly (the non-numeric requirement is documented) passes a numeric a:key to Concealer#AddLocal(), the char will revert to the configured one on a refresh.)